repositories: move BaseRepository to julee.repositories.base, remove julee.domain
BaseRepository is generic framework infrastructure — nothing CEAP-specific
about it. With the CEAP content now in contrib/ceap, julee/domain/ was
left holding only base.py and empty __init__ files. Move it to
julee.repositories.base alongside the concrete implementations that
depend on it, and remove the now-empty domain package entirely.
