+++ src/julee/contrib/ceap/domain/repositories/assembly_specification.py
@@ -34,7 +34,7 @@ from typing import Protocol, runtime_checkable
 from julee.contrib.ceap.domain.models.assembly_specification import (
     AssemblySpecification,
 )
-from julee.domain.repositories.base import BaseRepository
+from julee.repositories.base import BaseRepository
 
 
 @runtime_checkable
