+++ src/julee/contrib/ceap/domain/repositories/document_policy_validation.py
@@ -32,7 +32,7 @@ stubs that delegate to activities for durability and proper error handling.
 from typing import Protocol, runtime_checkable
 
 from julee.contrib.ceap.domain.models.policy import DocumentPolicyValidation
-from julee.domain.repositories.base import BaseRepository
+from julee.repositories.base import BaseRepository
 
 
 @runtime_checkable
