--- src/julee/domain/models/__init__.py
@@ -1,5 +1,0 @@
-"""
-Domain models for julee.
-
-CEAP domain models have moved to julee.contrib.ceap.domain.models.
-"""
