--- src/julee/domain/repositories/__init__.py
@@ -1,12 +1,0 @@
-"""
-Repository protocols for julee domain.
-
-CEAP repository protocols have moved to julee.contrib.ceap.domain.repositories.
-BaseRepository remains here as generic framework infrastructure.
-"""
-
-from .base import BaseRepository
-
-__all__ = [
-    "BaseRepository",
-]
