--- src/julee/domain/use_cases/__init__.py
@@ -1,5 +1,0 @@
-"""
-Use cases for julee domain.
-
-CEAP use cases have moved to julee.contrib.ceap.use_cases.
-"""
